Fix LZB decompressor: offset byte must be signed (s8), not unsigned
The decompiler used lbu (load byte unsigned) but the original N64
assembly likely uses lb (load byte signed). This one-instruction
difference changes all offset calculations where byte >= 128.

Critical evidence: with unsigned byte, the end marker (offset=-1)
is mathematically unreachable — (iVar7-3)*256 + unsigned_byte can
never equal -1 for valid iVar7 values. With signed byte,
(3-3)*256 + (s8)0xFF = 0 + (-1) = -1, correctly triggering the
end marker.

This also changes back-reference distances for all bytes with bit 7
set, producing different (hopefully correct) decompressed output.

+++ src_pseudo/lzb.cpp
@@ -88,7 +88,7 @@ LAB_800aa478:
       LZB_CHECK_IN("offset-byte");
       if (0) fprintf(stderr, "[lzb] iter=%u: offset extra byte=0x%02x → iVar7=(%d-3)*256+%u=%d\n",
               iter, compDat[uVar6], iVar7, (u32)compDat[uVar6], (iVar7 + -3) * 0x100 + (s32)(u32)compDat[uVar6]);
-      iVar7 = (iVar7 + -3) * 0x100 + (u32)compDat[uVar6];
+      iVar7 = (iVar7 + -3) * 0x100 + (s32)(s8)compDat[uVar6];
       uVar6++;
       if (iVar7 == -1) {
         *outSize = iVar9;
